requirements.txt
bs4==0.0.1
lxml==4.5.0
mysql-connector==2.2.9
pandas==1.0.1
requests==2.22.0
SQLAlchemy==1.3.13
